Modified Resources and Providers creation
This is a crazy commit that modified the way we create the Resources
and Providers inside Cheffish.

Instead of creating them as:
```
class Chef::Resource::Blah
```

We now use:
```
class Chef
class Resource
class Blah
```

The same for the Providers.

The main reason for this is to open Chef class and consume what is
in there. In this case the `NOT_PASSED` constant!