fix completely object model insanity
Chef::Runner takes the Chef::RunContext as its argument.  This class is
supposed to be an analogue of Chef::Client but its injecting *itself*
into the Chef::Runner, it then has to proxy all the messages that the
Chef::Runner makes to to the Chef::RunContext which creates hugely tight
coupling between this class and the already tightly coupled Chef::Runner
and Chef::RunContext classes.