Avoid redefining the Chef::Resource#name method
In chef/chef#5606, Chef will now log deprecation warnings whenever
a resource defines a property that is already an instance_method
on that resource. If this happens, a user may be inadvertently
causing unexpected and difficult-to-troubleshoot behavior.

A number of the cheffish resources were defining a `name` property
which is already a property defined on the Chef::Resource base class.
This change stops defining the `name` property for each of these
resources and instead defines something unique.

In addition, the `chef_mirror` resource defined a `freeze` property
which is already an instance_method on `Object` instances. I have
changed that to be `freeze_on_upload` and defined a method to log
a deprecation warning while providing backward compatibility.

Signed-off-by: Adam Leff <adam@leff.co>