Fixes for Chef-17
Updates for the removal of the chef-fs parallelizer.

This will be totally broken until we merge that PR into
chef-client itself, this is necessary to get that PR
green though and testing will continue there.

This drops support for chef <= 16, future versions of
cheffish will likely be brought into the chef-client
repo itself

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>